fixed compilation on WIN32 (MSVC12) in TransportTCP_WIN32.h
- fixes error:
"C:\Users\hperrey\Desktop\eudaq\build\INSTALL.vcxproj" (default target) (1) ->
"C:\Users\hperrey\Desktop\eudaq\build\ALL_BUILD.vcxproj" (default target) (2) ->
"C:\Users\hperrey\Desktop\eudaq\build\main\exe\ClusterExtractor.exe.vcxproj" (default target) (3) ->
"C:\Users\hperrey\Desktop\eudaq\build\main\lib\EUDAQ.vcxproj" (default target) (4) ->
(ClCompile target) ->
  C:/Users/hperrey/Desktop/eudaq/./main/include\eudaq/TransportTCP_WIN32.h(95): error C2664: 'int ioctlsocket(SOCKET,long,u_long *)' : cann
ot convert argument 3 from 'uint32_t *' to 'u_long *' [C:\Users\hperrey\Desktop\eudaq\build\main\lib\EUDAQ.vcxproj]

    13 Warning(s)

+++ main/include/eudaq/TransportTCP_WIN32.h
@@ -91,7 +91,7 @@ namespace eudaq {
     }
 
     static void setup_socket(SOCKET sock) {
-      uint32_t one = 1;
+      unsigned long one = 1;
       ioctlsocket(sock, FIONBIO, &one);
     }
